# Conflicts:
#	server.py
#	templates/display_question/question_answers.html

+++ server.py
@@ -61,7 +61,6 @@ def route_index():
     # if 'username' in session:
     #     username = session['username']
     #     return render_template('home/index.html', sorted_questions=sorted_questions, user=username)
-
     return render_template('home/index.html', sorted_questions=sorted_questions)
 
 
@@ -113,33 +112,42 @@ def display_question_and_answers(question_id):
     answers = data_manager.get_answers_for_question(question_id)
     tags = data_manager.get_tags_for_question(question_id)
     comments = data_manager.get_all_comments(question_id)
+    user_id = session.get('user_id', False)
+
     return render_template('display_question/question_display.html', question=question, tags=tags,
-                           answers=answers, question_ids=question_ids, comments=comments)
+                           answers=answers, question_ids=question_ids, comments=comments, user_id=user_id)
 
 
 @app.route('/question/<question_id>/vote', methods=['POST'])
 def route_vote(question_id):
     vote_option, message_id, message_type = request.form['vote'].split(',')
     data_manager.handle_votes(vote_option, message_id, message_type)
+    data_manager.handle_user_reputation(vote_option, message_id, message_type)
 
     # the code=307 argument ensures that the request type (POST) is preserved after redirection
     # so that the view number of the question doesn't increase after voting
     return redirect(url_for('display_question_and_answers', question_id=question_id), code=307)
 
 
+@app.route('/question/<question_id>/<answer_id>/accepted_answer', methods=['GET'])
+def route_accepted_answer(question_id, answer_id):
+    data_manager.handle_accepted_answer(question_id, answer_id)
+    data_manager.handle_user_reputation('accepted_answer', answer_id)
+
+    # the code=307 argument ensures that the request type (POST) is preserved after redirection
+    # so that the view number of the question doesn't increase after voting
+    return redirect(url_for('display_question_and_answers', question_id=question_id))
+
+
 @app.route('/question/<question_id>/edit', methods=['GET', 'POST'])
 def route_edit_question(question_id):
+    if data_manager.question_belongs_to_user(session.get('username'), question_id):
+        if request.method == 'GET':
+            question_data = data_manager.get_single_question(question_id)
+            return render_template('database_ops/add-question.html', question_data=question_data)
 
-    if request.method == 'GET':
-        if 'user_id' in session:
-            user_id_for_question = data_manager.get_user_id_for_question(question_id)
-            if session['user_id'] == user_id_for_question:
-                question_data = data_manager.get_single_question(question_id)
-                return render_template('database_ops/add-question.html', question_data=question_data)
-        return redirect(url_for('display_question_and_answers', question_id=question_id))
-
-    user_inputs_for_question = request.form.to_dict()
-    data_manager.update_entry('question', question_id, user_inputs_for_question)
+        user_inputs_for_question = request.form.to_dict()
+        data_manager.update_entry('question', question_id, user_inputs_for_question)
 
     return redirect(url_for('display_question_and_answers', question_id=question_id), code=307)
 
@@ -164,17 +172,16 @@ def route_new_answer(question_id):
 
 @app.route('/question/<question_id>/delete')
 def route_delete_question(question_id):
-    if 'user_id' in session:
-        user_id_for_question = data_manager.get_user_id_for_question(question_id)
-        if session['user_id'] == user_id_for_question:
-            data_manager.delete_question(question_id)
-            return redirect(url_for('route_index'))
+    if data_manager.question_belongs_to_user(session.get('username'), question_id):
+        data_manager.delete_question(question_id)
+        return redirect(url_for('route_index'))
     return redirect(url_for('display_question_and_answers', question_id=question_id))
 
 
 @app.route('/question/<question_id>/<answer_id>/delete')
 def route_delete_answer(question_id, answer_id):
-    data_manager.delete_answer(answer_id)
+    if data_manager.answer_belongs_to_user(session.get('username'), answer_id):
+        data_manager.delete_answer(answer_id)
     return redirect(url_for('display_question_and_answers', question_id=question_id))
 
 
@@ -184,11 +191,13 @@ def route_edit_answer(answer_id):
     question_id = answer_data.get('question_id')
     question_data = data_manager.get_single_entry('question', question_id)
 
-    if request.method == 'GET':
-        return render_template('database_ops/new_answer.html', answer=answer_data, question=question_data)
+    if data_manager.answer_belongs_to_user(session.get('username'), answer_id):
 
-    user_inputs_for_answer = request.form.to_dict()
-    data_manager.update_entry('answer', answer_id, user_inputs_for_answer)
+        if request.method == 'GET':
+            return render_template('database_ops/new_answer.html', answer=answer_data, question=question_data)
+
+        user_inputs_for_answer = request.form.to_dict()
+        data_manager.update_entry('answer', answer_id, user_inputs_for_answer)
 
     return redirect(url_for('display_question_and_answers', question_id=question_id), code=307)
 
@@ -196,27 +205,21 @@ def route_edit_answer(answer_id):
 @app.route('/question/<question_id>/new-tag', methods=["GET", "POST"])
 def route_new_tag(question_id):
     existing_tags = data_manager.get_existing_tags_for_question(question_id, )
+    if data_manager.question_belongs_to_user(session.get('username'), question_id):
+        return render_template('database_ops/new_tag.html', existing_tags=existing_tags)
 
     if request.method == "POST":
         new_tag = request.form.get('new_tag')
         existing_tag_id = request.form.get('existing_tag')
         data_manager.handle_tag(question_id, new_tag, existing_tag_id)
-        return redirect(url_for('display_question_and_answers', question_id=question_id), code=307)
 
-    if 'user_id' in session:
-        user_id_for_question = data_manager.get_user_id_for_question(question_id)
-        if session['user_id'] == user_id_for_question:
-            return render_template('database_ops/new_tag.html', existing_tags=existing_tags)
-
-    return redirect(url_for('display_question_and_answers', question_id=question_id))
+    return redirect(url_for('display_question_and_answers', question_id=question_id), code=307)
 
 
 @app.route('/question/<question_id>/tag/<tag_id>/delete')
 def route_delete_tag(question_id, tag_id):
-    if 'user_id' in session:
-        user_id_for_question = data_manager.get_user_id_for_question(question_id)
-        if session['user_id'] == user_id_for_question:
-            data_manager.delete_tag(question_id, tag_id)
+    if data_manager.question_belongs_to_user(session.get('username'), question_id):
+        data_manager.delete_tag(question_id, tag_id)
 
     return redirect(url_for('display_question_and_answers', question_id=question_id))
 
@@ -240,7 +243,8 @@ def route_add_comment_to_answer(question_id, answer_id):
     # After this process it redirects you to the specific page of the question.
 
     comment_message = request.form['message']
-    data_manager.insert_comment(comment_message, question_id, answer_id=answer_id)
+    user_id = session['user_id']
+    data_manager.insert_comment(comment_message, question_id, answer_id=answer_id, user_id=user_id)
     return redirect(url_for('display_question_and_answers', question_id=question_id), code=307)
 
 
@@ -250,8 +254,10 @@ def route_add_comment_to_question(question_id):
         question = data_manager.get_single_entry('question', question_id)
         return render_template('database_ops/new_comment.html', answer_by_id=question)
 
+    user_id = session['user_id']
     comment_message = request.form['message']
-    data_manager.insert_comment(comment_message, question_id)
+    data_manager.insert_comment(comment_message, question_id, user_id)
+
     return redirect(url_for('display_question_and_answers', question_id=question_id), code=307)
 
 
@@ -271,18 +277,19 @@ def route_tags():
 @app.route('/comment/<comment_id>/delete', methods=["GET", "POST"])
 def route_delete_comment(comment_id):
     comment = data_manager.get_single_entry('comment', comment_id)
-    answer_id_of_comment = comment['answer_id']
-    question_id_of_comment = comment['question_id']
+    if data_manager.comment_belongs_to_user(session.get('username'), comment_id):
+        answer_id_of_comment = comment['answer_id']
+        question_id_of_comment = comment['question_id']
 
-    if request.method == "GET":
-        if answer_id_of_comment:
-            answer_data = data_manager.get_single_entry('answer', answer_id_of_comment)
-            return render_template('database_ops/delete_comment.html', answer=answer_data['message'], comment=comment)
-        question_data = data_manager.get_single_entry('question', question_id_of_comment)
-        return render_template('database_ops/delete_comment.html', question=question_data, comment=comment)
+        if request.method == "GET":
+            if answer_id_of_comment:
+                answer_data = data_manager.get_single_entry('answer', answer_id_of_comment)
+                return render_template('database_ops/delete_comment.html', answer=answer_data['message'], comment=comment)
+            question_data = data_manager.get_single_entry('question', question_id_of_comment)
+            return render_template('database_ops/delete_comment.html', question=question_data, comment=comment)
 
-    if request.form['delete-button'] == 'Yes':
-        data_manager.delete_comment(comment_id)
+        if request.form['delete-button'] == 'Yes':
+            data_manager.delete_comment(comment_id)
 
     return redirect(url_for('display_question_and_answers', question_id=comment['question_id']), code=307)
 
@@ -291,17 +298,19 @@ def route_delete_comment(comment_id):
 def route_edit_comment(comment_id):
     comment_data = data_manager.get_single_entry('comment', comment_id)
     question_id = comment_data.get('question_id')
-    question_data = data_manager.get_single_entry('question', question_id)
-    answer_data = None
-    if comment_data['answer_id']:
-        answer_id = comment_data['answer_id']
-        answer_data = data_manager.get_single_entry('answer', answer_id)
 
-    if request.method == 'GET':
-        return render_template('database_ops/new_comment.html', comment=comment_data, answer=answer_data, question=question_data)
+    if data_manager.comment_belongs_to_user(session.get('username'), comment_id):
+        question_data = data_manager.get_single_entry('question', question_id)
+        answer_data = None
+        if comment_data['answer_id']:
+            answer_id = comment_data['answer_id']
+            answer_data = data_manager.get_single_entry('answer', answer_id)
 
-    new_comment_message = request.form['message']
-    data_manager.update_comment_message(comment_data, new_comment_message)
+        if request.method == 'GET':
+            return render_template('database_ops/new_comment.html', comment=comment_data, answer=answer_data, question=question_data)
+
+        new_comment_message = request.form['message']
+        data_manager.update_comment_message(comment_data, new_comment_message)
 
     return redirect(url_for('display_question_and_answers', question_id=question_id), code=307)
 
@@ -320,6 +329,16 @@ def record_user(user_data):
     username_is_unique = data_manager.is_username_unique(user_data['username'])
     if username_is_unique:
         data_manager.insert_user(user_data)
+
+
+@app.route('/user/<user_id>')
+def route_user_page(user_id):
+    user_id = int(user_id)
+    if not ('user_id' in session and session['user_id'] == user_id):
+        return redirect('/')
+
+    user_data = data_manager.get_user_data_for_user_page(session['user_id'], session['username'])
+    return render_template('user_page/main.html', user_data=user_data)
 
 
 if __name__ == '__main__':
